fix: move code to base/participants and check origin
